refactor(core): lift access managers into unifi_core.access.managers
8 manager modules moved with backwards-compatible shims at the old paths.
Cross-manager and lifted-shared imports rewritten. Shims removed in Task 17.

The door_manager shim re-exports the private _LOCATIONS_EXPAND symbol used
by unit tests. Updated test_connection_manager patches to target the new
unifi_core.access.managers.connection_manager module path so aiohttp
patches resolve in the right namespace.
